Deep elf archers.
They are to master archers as knights are to blademasters - a mid-tier
threat with some support spellcasting on the side. In this case, it's
more of an AM theme.

Contains some adjustments to Blink Other to make it more useful in this
case - in particular it has a very high chance (guarantee?) to place the
victim in LOS of the caster. This touches a codepath that may also
impact jumping spider and Disjunction behaviour slightly.

Additional spell set ideas welcome.

TODO: place these sometimes in vaults that place elves.