Monster spell: Aura of Brilliance, for deep elf high priests.
Fixes the priest in place and grants nearby wizard allies of the priest
the ability to cast spells more often (they get a reroll if they rolled
higher than their total spell weight) with a 5 HD boost in their spell
power. The boost lasts for a turn or two after the wizard leaves LOS of
the priest (including if the priest is killed).

Presently the effect lasts 200 aut or until there are no allies in LOS
that benefit from its effect. This could use tweaking.