Remove Blade (again); place "good" weapons in Elf:$ and others in Elf.
Blade in Elf has exactly the same problems as it did when it was its own
branch: players avoided it completely unless they were specifically
looking for something in the loot. Having it occupy a vault slot also
has the disadvantage of blocking off a large section of the level and
preventing other vaults from placing.

Redistributing the vault in this fashion solves the above problems, in
addition to providing us with another enemy to face in Elf that isn't a
"squishy elf".

This borrows some previous logic for placing an extra vault on Elf:1 to
also try to place an extra vault on Elf:2.