Add Doom Hounds
New enemy for Tartarus. Use a smitey hex that, if it beats MR,
causes more hounds (avg 2.5) to spawn just out of sight for the
next 8-12 turns. An escalation of the situation, vaguely like
Sentinel's Mark.

Questions:
1) Right now they're vulnerable to silence, which doesn't play
that great with Tartarus's silent spectres & Eresh herself. The
effect is already themed as partially mental - should it be made
entirely so?
2) How reasonable are their numbers? They're supposed to be about
2/3rds of a reaper, which might be a bit much in large packs...